requirements.txt
alembic==1.4.3
amqp==5.0.2
aniso8601==8.0.0
billiard==3.6.3.0
boto3==1.16.25
botocore==1.19.25
celery==5.0.2
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
flask-restful-swagger==0.20.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==5.0.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.24.1
passlib==1.7.4
prompt-toolkit==3.0.8
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
s3transfer==0.3.3
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
